requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
scikit-image==0.22.0
imagehash==4.3.1
google-generativeai==0.3.1
